+++ app/controllers/qbrick/pages_controller.rb
@@ -1,19 +1,20 @@
 module Qbrick
   class PagesController < ::ApplicationController
     respond_to :html
-    before_action :find_page_by_url, only: :show
+    before_action :set_locale
+    before_action :find_page_by_path, only: :show
 
     def index
       @search = params[:search]
       return if @search.blank?
+
       @pages = Qbrick::Page.unscoped.published.content_page.search(@search)
     end
 
     def show
       if redirect_page?
-        redirect_url = @page.redirect_url.sub(%r{\A\/+}, '') # remove all preceding slashes
         session[:qbrick_referrer] = @page.id
-        redirect_to "/#{redirect_url}"
+        redirect_to @page.redirect_url
       elsif @page.present?
         respond_with @page
       elsif @page.blank? && respond_to?(:handle_404)
@@ -30,14 +31,22 @@ module Qbrick
 
     private
 
+    def set_locale
+      new_locale = params[:locale] || session['frontend_locale'] || I18n.locale
+
+      return I18n.locale if I18n.locale == new_locale || !I18n.locale_available?(new_locale)
+
+      session['frontend_locale'] = new_locale.to_s
+      I18n.locale = new_locale
+    end
+    alias frontend_locale set_locale
+
     def redirect_page?
       @page.present? && @page.redirect? && @page.redirect_url.present?
     end
 
-    def find_page_by_url
-      url = locale.to_s
-      url += "/#{params[:url]}" if params[:url].present?
-      @page = Qbrick::Page.find_by_url(url)
+    def find_page_by_path
+      @page = Qbrick::Page.published.find_by_path params[:url].to_s
     end
   end
 end