+++ app/controllers/qbrick/sitemaps_controller.rb
@@ -3,9 +3,10 @@ module Qbrick
     def index
       last_page = Qbrick::Page.published.last
       return unless stale?(etag: last_page, last_modified: last_page.updated_at.utc)
+
       respond_to do |format|
         format.html
-        format.xml { @pages = Qbrick::Page.published }
+        format.xml { @pages = Qbrick::Page.all }
       end
     end
   end