+++ app/models/qbrick/page.rb
@@ -10,12 +10,12 @@ module Qbrick
     acts_as_brick_list
 
     translate :title, :page_title, :slug, :keywords, :description,
-              :body, :redirect_url, :url
+              :body, :redirect_url, :path, :published
 
     default_scope { order 'position ASC' }
 
-    scope :published, -> { where published: Qbrick::PublishState::PUBLISHED }
-    scope :translated, -> { where "url_#{I18n.locale.to_s.underscore} is not null" }
+    scope :published, -> { where locale_attr(:published) => Qbrick::PublishState::PUBLISHED }
+    scope :translated, -> { where.not locale_attr(:path) => [nil, ''], locale_attr(:title) => [nil, ''] }
 
     scope :content_page, -> { where page_type: Qbrick::PageType::CONTENT }
 
@@ -25,14 +25,16 @@ module Qbrick
         locale_attr(:page_type) => Qbrick::PageType::NAVIGATION)
     }
 
-    before_validation :create_slug, :create_url
-    after_save :update_child_urls
+    before_validation :create_slug, :create_path
+    after_save :update_child_paths
+    validate :remove_preceding_slashes, if: :redirect?
 
     validates :title, presence: true
     validates :slug, presence: true
     validates :redirect_url, presence: true, if: :redirect?
     validates :title, :slug, :keywords, :page_type, length: { maximum: 255 }
     validates :identifier, uniqueness: true, allow_blank: true
+    validate :slug_uniqueness
 
     class << self
       def flat_tree
@@ -55,14 +57,38 @@ module Qbrick
         find_by(identifier: identifier)
       end
 
-      def all_urls
-        url_columns = column_names.select { |col| col.start_with? 'url_' }
-        pluck(*url_columns).flatten.compact.sort.uniq.map { |r| "/#{r}" }
+      def all_paths
+        path_columns = column_names.select { |col| col.start_with? 'path_' }
+        pluck(*path_columns).flatten.compact.sort.uniq.map(&:path)
       end
+
+      def find_by_path(given_path)
+        find_by locale_attr(:path) => given_path.blank? ? '' : "/#{given_path.sub(%r{^/+}, '')}"
+      end
+    end # class methods
+
+    def slug_uniqueness
+      path_field = locale_attr :path
+      slug_field = locale_attr :slug
+      [slug_field, path_field].each do |field|
+        self.class.validators_on(field).map { |v| v.validate self }
+        return true if errors[field].present?
+      end
+
+      page_with_duplicated_paths = self.class.published.translated.where path_field => path
+      page_with_duplicated_paths = page_with_duplicated_paths.where.not id: id if persisted?
+      return true unless page_with_duplicated_paths.exists?
+
+      message = 'page ids: '
+      page_with_duplicated_paths.order(:id).pluck(:id).each do |id|
+        message << "<a href=\"#{edit_cms_page_path id}#page-metadata\" target=\"_blank\">#{id}</a>, "
+      end
+      message = I18n.t 'activerecord.errors.models.qbrick/page.attributes.slug.duplicated_slug', append: " (#{message.sub(/, $/, '')})"
+      errors.add :slug, message.html_safe
     end
 
     def without_self
-      self.class.where 'id != ?', id
+      self.class.where.not id: id
     end
 
     def published?
@@ -77,6 +103,29 @@ module Qbrick
       page_type == Qbrick::PageType::REDIRECT || page_type == Qbrick::PageType::CUSTOM
     end
 
+    def internal_redirect?
+      return false unless redirect?
+
+      scheme = URI.parse(redirect_url).scheme
+      return true if scheme.nil?
+
+      internal_redirect = Qbrick::Engine.hosts.find do |h|
+        URI.parse("#{scheme}://#{h}").route_to(redirect_url).host.nil?
+      end
+
+      internal_redirect.present?
+    end
+
+    def external_redirect?
+      redirect? && !internal_redirect?
+    end
+
+    def remove_preceding_slashes
+      return if redirect_url.blank?
+
+      redirect_url.sub!(%r{^/+}, '/')
+    end
+
     def navigation?
       page_type == Qbrick::PageType::NAVIGATION
     end
@@ -86,18 +135,18 @@ module Qbrick
     end
 
     def translated?
-      url.present? && title.present? && slug.present?
+      title.present? && slug.present?
     end
 
     def translated_to?(raw_locale)
       locale = raw_locale.to_s.underscore
-      send("url_#{locale}").present? && send("title_#{locale}").present? && send("slug_#{locale}").present?
+      send("title_#{locale}").present? && send("slug_#{locale}").present?
     end
 
     def translated_link_for(locale)
       if translated_to? locale
         I18n.with_locale locale do
-          url_with_locale
+          path_with_prefixed_locale
         end
       else
         Qbrick::Page.roots.first.link
@@ -106,45 +155,48 @@ module Qbrick
 
     def link
       if bricks.count == 0 && children.count > 0
-        children.first.link
+        children.published.first.link
       else
-        url_with_locale
+        path_with_prefixed_locale
       end
     end
 
-    # TODO: needs naming and routing refactoring (url/locale/path/slug)
     def path_segments
       paths = parent.present? ? parent.path_segments : []
       paths << slug unless navigation?
       paths
     end
 
-    def url_without_locale
-      path_segments.join('/')
+    def path_with_prefixed_locale(locale = I18n.locale)
+      "/#{locale}#{send self.class.attr_name_for_locale(:path, locale)}"
     end
 
-    def url_with_locale
+    def url
+      URI::HTTP.build([nil, Qbrick::Engine.host, Qbrick::Engine.port, path_with_prefixed_locale, nil, nil]).tap do |url|
+        url.scheme = Qbrick::Engine.scheme
+      end
+    end
+
+    def create_path
       opts = { locale: I18n.locale }
-      url = url_without_locale
-      opts[:url] = url if url.present?
-      page_path(opts)
-    end
+      path = path_segments.join '/'
+      opts[:url] = path if path.present?
 
-    def create_url
-      self.url = url_with_locale[1..-1]
+      self.path = page_path(opts).sub(%r{^/#{I18n.locale}}, '')
     end
 
     def create_slug
       if title.present? && slug.blank?
         self.slug = title.downcase.parameterize
       elsif slug.present?
-        self.slug = slug.downcase
+        self.slug = slug.downcase.parameterize
       end
     end
 
-    def update_child_urls
-      return unless children.any?
-      children.each { |child| child.update_attribute(:url, child.create_url) }
+    def update_child_paths
+      children.each do |child|
+        child.update_attribute :path, child.create_path
+      end
     end
 
     def nesting_name
@@ -170,11 +222,7 @@ module Qbrick
     end
 
     def as_json
-      {}.tap do |json|
-        json['title'] = send("title_#{I18n.locale.to_s.underscore}")
-        json['pretty_url'] = '/' + send("url_#{I18n.locale.to_s.underscore}")
-        json['url'] = "/pages/#{id}"
-      end
+      { 'title' => title, 'pretty_url' => path, 'url' => "/pages/#{id}" }
     end
 
     def clear_bricks_for_locale(locale)