+++ CHANGELOG.md
@@ -1,5 +1,54 @@
 # CHANGELOG
 
+## 2.6.10 - 2015-08-11
+
+* move to [`ckeditor` gem](https://github.com/galetahub/ckeditor) again
+* Page#link of an empty page will now show the first PUBLISHED child
+* Qbrick::Engine.hosts won't return empty values anymore
+
+## 2.6.9 - 2015-07-15
+
+* make translated scope consistent to the conditions in #translated?
+
+## 2.6.8 - 2015-07-14
+
+* validate uniqueness of slugs just against published pages
+
+## 2.6.7 - 2015-06-30
+
+* Add url method to get an url to the page
+* Add ability to add scheme, host and port to your app config
+
+## 2.6.6 - 2015-06-26
+
+* Also parameterize slugs that are explicit added
+
+## 2.6.5 - 2015-06-26
+
+* Add check on duplicated page paths
+
+## 2.6.3 - 2015-06-10
+
+* Simplify check on translated pages
+
+## 2.6.2 - 2015-06-10
+
+* handle internal and external redirects
+
+## 2.6.1 - 2015-06-03
+
+* url is now path and contains a leading slash
+* url_without_locale is now path_with_prefixed_locale and contains a leading slash as well
+* visibility bugfix for sitemaps
+* removed various deprecation warnings
+* refactored spec/controllers/qbrick/pages_controller_spec.rb
+
+## 2.6.0 - 2015-05-20
+
+* Enable publishing by locale
+* Fix page visibility bug
+* Fix locale switch issue: Switching cms backend affected frontend locale (inverse case of issue #10)
+
 ## 2.5.2 - 2015-05-08
 
 * Speed up brick cloning